Project 1 initial analysis and visualisation
The aim of this project is to gain an initial insight into the data set we will be using throughout the subject. This will be achieved through performing an initial analysis, along with a visualisation of the results. The data set we will be using throughout will be the New York City Taxi and Limousine Service Trip Record Data. The